+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2020-11-10
+- Add SHAP RFECV for features elimination
+
+## [1.3.0] - 2020-11-05
+- Add SHAP Model Inspector with docs and tests
+
 ## [1.2.0] - 2020-09-30
 - Add resemblance model, with SHAP based importance
 - Improve the docs for resemblance model